+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Split long description into short and long description fields.
+
 ### Fixed
 
 - Fix `redefine 'l' shorthand in "makefile" flagset` error.